Fix rounding, float precision, and type safety (#135)
* Fix rounding and float precision in plugins

- Use std::llround for time-to-frames conversions in node_control instead
  of truncating casts, avoiding off-by-one frame errors
- Use float arithmetic for Hermite interpolation in disksampler/sampler
  (interpolation coefficient x does not need double precision)
- Use if constexpr for compile-time wrapPhase branch in both samplers
- Use sinf in sine.c to match float output type
- Fix implicit narrowing in soundfile_api_dummy distribution/struct fields

* Fix type safety in engine and clean up build

- Use Methcla_PortCount for port index parameters in SynthDef, Synth, and
  EngineImpl instead of size_t, matching the plugin API type
- Add explicit casts for sampleOffset, sampleRate, and OSC integer fields
- Rename Loader.cpp to LoaderPosix.cpp to make the platform explicit
- Remove empty Resource.cpp (only contained a using directive)
- Drop _WIN32 aligned-memory path from Memory.cpp

+++ include/methcla/engine.hpp
@@ -867,7 +867,7 @@ namespace Methcla {
                 .openMessage("/node/set", 3)
                 .int32(node.id())
                 .int32(static_cast<int32_t>(index))
-                .float32(value)
+                .float32(static_cast<float>(value))
                 .closeMessage();
         }
